+++ nodeconductor/core/authentication.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 from django.conf import settings
+from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 import rest_framework.authentication
 from rest_framework import exceptions
@@ -23,6 +24,23 @@ class TokenAuthentication(rest_framework.authentication.TokenAuthentication):
         if not auth:
             auth = request.query_params.get(TOKEN_KEY, '')
         return auth
+
+    def authenticate_credentials(self, key):
+        try:
+            token = self.model.objects.select_related('user').get(key=key)
+        except self.model.DoesNotExist:
+            raise exceptions.AuthenticationFailed(_('Invalid token.'))
+
+        if not token.user.is_active:
+            raise exceptions.AuthenticationFailed(_('User inactive or deleted.'))
+
+        if token.created < timezone.now() - timezone.timedelta(hours=1):
+            raise exceptions.AuthenticationFailed(_('Token has expired.'))
+        else:
+            token.created = timezone.now()
+            token.save()
+
+        return token.user, token
 
     def authenticate(self, request):
         auth = self.get_authorization_value(request).split()