+++ nodeconductor/cost_tracking/views.py
@@ -1,9 +1,14 @@
 from __future__ import unicode_literals
 
+from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import ObjectDoesNotExist
+from django.db.models import Prefetch
+
 from rest_framework import viewsets, permissions, exceptions, decorators, response, status
 
 from nodeconductor.core.filters import DjangoMappingFilterBackend
 from nodeconductor.cost_tracking import models, serializers, filters
+from nodeconductor.structure import SupportedServices
 from nodeconductor.structure import models as structure_models
 from nodeconductor.structure.filters import ScopeTypeFilterBackend
 
@@ -183,7 +188,11 @@ class PriceListItemViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
     def list(self, request, *args, **kwargs):
         """
         To get a list of price list items, run **GET** against */api/price-list-items/* as an authenticated user.
+        """
+        return super(PriceListItemViewSet, self).list(request, *args, **kwargs)
 
+    def create(self, request, *args, **kwargs):
+        """
         Run **POST** request against */api/price-list-items/* to create new price list item.
         Customer owner and staff can create price items.
 
@@ -199,23 +208,27 @@ class PriceListItemViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
 
             {
                 "units": "per month",
-                "key": "test_key",
                 "value": 100,
                 "service": "http://example.com/api/oracle/d4060812ca5d4de390e0d7a5062d99f6/",
-                "item_type": "storage"
+                "default_price_list_item": "http://example.com/api/default-price-list-items/349d11e28f634f48866089e41c6f71f1/"
             }
         """
-        return super(PriceListItemViewSet, self).list(request, *args, **kwargs)
+        return super(PriceListItemViewSet, self).create(request, *args, **kwargs)
 
-    def retrieve(self, request, *args, **kwargs):
+    def update(self, request, *args, **kwargs):
         """
         Run **PATCH** request against */api/price-list-items/<uuid>/* to update price list item.
-        Only value and units can be updated. Customer owner and staff can update price items.
-
-        Run **DELETE** request against */api/price-list-items/<uuid>/* to delete price list item.
-        Customer owner and staff can delete price items.
+        Only item_type, key value and units can be updated.
+        Only customer owner and staff can update price items.
         """
-        return super(PriceListItemViewSet, self).retrieve(request, *args, **kwargs)
+        return super(PriceListItemViewSet, self).update(request, *args, **kwargs)
+
+    def destroy(self, request, *args, **kwargs):
+        """
+        Run **DELETE** request against */api/price-list-items/<uuid>/* to delete price list item.
+        Only customer owner and staff can delete price items.
+        """
+        return super(PriceListItemViewSet, self).destroy(request, *args, **kwargs)
 
     def initial(self, request, *args, **kwargs):
         if self.action in ('partial_update', 'update', 'destroy'):
@@ -252,3 +265,57 @@ class DefaultPriceListItemViewSet(viewsets.ReadOnlyModelViewSet):
          - ?resource_type=<string> resource type, for example: 'OpenStack.Instance, 'Oracle.Database')
         """
         return super(DefaultPriceListItemViewSet, self).list(request, *args, **kwargs)
+
+
+class MergedPriceListItemViewSet(viewsets.ReadOnlyModelViewSet):
+    lookup_field = 'uuid'
+    permission_classes = (permissions.IsAuthenticated,)
+    filter_class = filters.DefaultPriceListItemFilter
+    filter_backends = (DjangoMappingFilterBackend,)
+    serializer_class = serializers.MergedPriceListItemSerializer
+
+    def list(self, request, *args, **kwargs):
+        """
+        To get a list of price list items, run **GET** against */api/merged-price-list-items/*
+        as authenticated user.
+
+        If service is not specified default price list items are displayed.
+        Otherwise service specific price list items are displayed.
+        In this case rendered object contains {"is_manually_input": true}
+
+        In order to specify service pass query parameters:
+        - service_type (Azure, OpenStack etc.)
+        - service_uuid
+
+        Example URL: http://example.com/api/merged-price-list-items/?service_type=Azure&service_uuid=cb658b491f3644a092dd223e894319be
+        """
+        return super(MergedPriceListItemViewSet, self).list(request, *args, **kwargs)
+
+    def get_queryset(self):
+        queryset = models.DefaultPriceListItem.objects.all()
+        service = self._find_service()
+        if service:
+            # Filter items by resource type
+            resources = SupportedServices.get_related_models(service)['resources']
+            content_types = ContentType.objects.get_for_models(*resources).values()
+            queryset = queryset.filter(resource_content_type__in=content_types)
+
+            # Attach service-specific items
+            price_list_items = models.PriceListItem.objects.filter(service=service)
+            prefetch = Prefetch('pricelistitem_set', queryset=price_list_items, to_attr='service_item')
+            queryset = queryset.prefetch_related(prefetch)
+        return queryset
+
+    def _find_service(self):
+        service_type = self.request.query_params.get('service_type')
+        service_uuid = self.request.query_params.get('service_uuid')
+        if not service_type or not service_uuid:
+            return
+        rows = SupportedServices.get_service_models()
+        if service_type not in rows:
+            return
+        service_class = rows.get(service_type)['service']
+        try:
+            return service_class.objects.get(uuid=service_uuid)
+        except ObjectDoesNotExist:
+            return None