+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Release NEXT
 
 - Added expiration time to authorization tokens.
 - Fix filtering events by scope_type and time range.
+- Implemented custom provider pricing configuration.
 
 Release 0.96.0
 --------------